refactor(goldgetters): move trpc functionality to data-access and server libs
