+++ apps/goldgetters/pages/_app.tsx
@@ -5,9 +5,7 @@ import { Header } from '@pellegrims/goldgetters/ui/organisms';
 import { PageTemplate } from '@pellegrims/goldgetters/ui/templates';
 import logo from '../public/logo.png';
 import { AppType } from 'next/dist/shared/lib/utils';
-import { withTRPC } from '@trpc/next';
-import { AppRouter } from './api/trpc/[trpc]';
-import { getBaseUrl } from '../utils/url';
+import { withGoldgettersTRPC } from '@pellegrims/goldgetters/data-access';
 
 const headerNavLinks: { text: string; href: string }[] = [
   { text: 'Nieuws', href: '/news' },
@@ -31,6 +29,4 @@ const GoldgettersApp: AppType = ({ Component, pageProps }) => (
   </>
 );
 
-export default withTRPC<AppRouter>({
-  config: () => ({ url: `${getBaseUrl()}/api/trpc` }),
-})(GoldgettersApp);
+export default withGoldgettersTRPC(GoldgettersApp);
