--- apps/goldgetters/utils/trpc.ts
@@ -1,4 +1,0 @@
-import { createReactQueryHooks } from '@trpc/react';
-import type { AppRouter } from '../pages/api/trpc/[trpc]';
-
-export const trpc = createReactQueryHooks<AppRouter>();
