+++ libs/goldgetters/data-access/README.md
@@ -1,0 +1,11 @@
+# goldgetters-data-access
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test goldgetters-data-access` to execute the unit tests via [Jest](https://jestjs.io).
+
+## Running lint
+
+Run `nx lint goldgetters-data-access` to execute the lint via [ESLint](https://eslint.org/).
