+++ libs/goldgetters/data-access/src/index.ts
@@ -1,0 +1,1 @@
+export { trpc, withGoldgettersTRPC } from './lib/trpc';
