+++ libs/goldgetters/data-access/src/lib/trpc.ts
@@ -1,0 +1,19 @@
+import { createReactQueryHooks } from '@trpc/react';
+import type { AppRouter } from '@pellegrims/goldgetters/server';
+import { withTRPC } from '@trpc/next';
+
+export const trpc = createReactQueryHooks<AppRouter>();
+
+export const withGoldgettersTRPC = withTRPC<AppRouter>({
+  config: () => ({ url: `${getBaseUrl()}/api/trpc` }),
+});
+
+const getBaseUrl = () => {
+  if (typeof window !== 'undefined') {
+    return '';
+  }
+  if (process.env.VERCEL_URL) {
+    return `https://${process.env.VERCEL_URL}`;
+  }
+  return `http://localhost:${process.env.PORT ?? 3000}`;
+};
