+++ libs/goldgetters/server/README.md
@@ -1,0 +1,11 @@
+# goldgetters-server
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test goldgetters-server` to execute the unit tests via [Jest](https://jestjs.io).
+
+## Running lint
+
+Run `nx lint goldgetters-server` to execute the lint via [ESLint](https://eslint.org/).
