+++ libs/goldgetters/server/src/index.ts
@@ -1,0 +1,2 @@
+export { nextApiHandler } from './lib/server';
+export type { AppRouter } from './lib/router';
