+++ libs/goldgetters/server/src/lib/server.ts
@@ -1,0 +1,7 @@
+import * as trpcNext from '@trpc/server/adapters/next';
+import { appRouter } from './router';
+
+export const nextApiHandler = trpcNext.createNextApiHandler({
+  router: appRouter,
+  createContext: () => null,
+});
