+++ tsconfig.base.json
@@ -16,7 +16,13 @@
     "baseUrl": ".",
     "paths": {
       "@pellegrims/eslint-config": ["libs/eslint-config/src/index.ts"],
+      "@pellegrims/goldgetters/data-access": [
+        "libs/goldgetters/data-access/src/index.ts"
+      ],
       "@pellegrims/goldgetters/db": ["libs/goldgetters/db/src/index.ts"],
+      "@pellegrims/goldgetters/server": [
+        "libs/goldgetters/server/src/index.ts"
+      ],
       "@pellegrims/goldgetters/ui/atoms": [
         "libs/goldgetters/ui/atoms/src/index.ts"
       ],
